refactor: Export 로직 PlatformHandler로 이식
