+++ ShareCode/ShareCode/src/Platform/MacOSPlatformHandler.cs
@@ -1,15 +1,10 @@
-﻿using System.Collections.ObjectModel;
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using UIKit;
 
 namespace ShareCode.src.Platform
 {
     public class MacOSPlatformHandler : PlatformHandler
     {
-        public override async Task OnExportButtonClicked(ContentPage page, ObservableCollection<FileItem> fileList)
-        {
-        }
-        
         public override void ConfigureWindow()
         {
             if (UIApplication.SharedApplication.ConnectedScenes.FirstOrDefault<object>() is not UIWindowScene windowScene) return;
